build error fixes

+++ MPC.Webstore/Areas/WebstoreApi/Controllers/RFQController.cs
@@ -136,7 +136,7 @@ namespace MPC.Webstore.Areas.WebstoreApi.Controllers
             NewInqury.Status = 25; // unproccessed status of inquiry
             int iMaxFileSize = 50000000;
             long result = _ItemService.AddInquiryAndItems(NewInqury, FillItems(InquiryItemDeliveryDate1, InquiryItemDeliveryDate2, InquiryItemDeliveryDate3, InquiryItemTitle1, InquiryItemNotes1, InquiryItemTitle2, InquiryItemNotes2, InquiryItemTitle3, InquiryItemNotes3, Convert.ToInt32(hfNoOfRec)));
-            int InquiryId = result;
+            long InquiryId = result;
 
             if (Request != null)
             {
